+++ pages/detail/index.tsx
@@ -5,20 +5,20 @@ import { styled } from 'stitches.config';
 import { useMutationDeleteMeeting, useMutationPostApplication, useQueryGetMeeting } from '@api/meeting/hooks';
 import { useRouter } from 'next/router';
 import Loader from '@components/loader/Loader';
-import InformationPanel from '@components/page/meetingDetail/InformationPanel';
+import InformationPanel from '@components/page/meetingDetail/information/InformationPanel';
 import { Tab } from '@headlessui/react';
-import FeedPanel from '@components/page/meetingDetail/FeedPanel';
+import FeedPanel from '@components/page/meetingDetail/feed/FeedPanel';
 import { Fragment, useState } from 'react';
 
 const enum SelectedTab {
-  INFORMATION,
   FEED,
+  INFORMATION,
 }
 
 const DetailPage = () => {
   const router = useRouter();
   const id = router.query.id as string;
-  const [selectedIndex, setSelectedIndex] = useState(SelectedTab.INFORMATION);
+  const [selectedIndex, setSelectedIndex] = useState(SelectedTab.FEED);
   const { data: detailData } = useQueryGetMeeting({ params: { id } });
   const { mutate: mutateDeleteMeeting } = useMutationDeleteMeeting({});
   const { mutate: mutatePostApplication } = useMutationPostApplication({});
@@ -38,18 +38,18 @@ const DetailPage = () => {
       <Tab.Group selectedIndex={selectedIndex} onChange={index => setSelectedIndex(index)}>
         <STabList>
           <Tab as={Fragment}>
-            <STabButton isSelected={selectedIndex === SelectedTab.INFORMATION}>모임 안내</STabButton>
+            <STabButton isSelected={selectedIndex === SelectedTab.FEED}>피드</STabButton>
           </Tab>
           <Tab as={Fragment}>
-            <STabButton isSelected={selectedIndex === SelectedTab.FEED}>피드</STabButton>
+            <STabButton isSelected={selectedIndex === SelectedTab.INFORMATION}>모임 안내</STabButton>
           </Tab>
         </STabList>
         <Tab.Panels>
           <Tab.Panel>
-            <InformationPanel detailData={detailData} />
+            <FeedPanel isMember={detailData?.approved || detailData?.host} />
           </Tab.Panel>
           <Tab.Panel>
-            <FeedPanel />
+            <InformationPanel detailData={detailData} />
           </Tab.Panel>
         </Tab.Panels>
       </Tab.Group>
