+++ src/components/page/meetingDetail/information/InformationPanel.tsx
